move universal editing functionality (button, undo/redo) from neonView to EditControls
